JS event clash detection :) Finished mapping events from PHP into JS :D ...might not be the most efficient thing ever so could be interesting to see how long it takes for the JS to parse all the events...
As for the view, works fine in FF, surprisingly is only a few pixels out in IE7, yet to test IE6 and other less common browsers. Any help with this is greatly appreciated. Also needs thoroughly testing with as many event combos as possible to make sure its working as expected.