Fixed rare bug reported by Food editor. Also now allowing food slideshow to get images from the gallery.
The bug reported by the food editor was an sql error that shows up if you add a photo to a slideshow that is already there. This was a rare bug as currently you could only add images by uploading a new one. So she must have uploaded it successfully, then refreshed the page forcing it to try and re-enter it into the slideshow.